Extract openDetails function
