+++ src/pages/css-props/index.js
@@ -1,23 +1,6 @@
 import { useEffect, useMemo } from "react";
 import cssPropsHtml from "./_cssPropsHtml";
-
-function openDetails() {
-  const hash = window.location.hash;
-  if (hash) {
-    const id = hash.substring(1); // Remove the '#' from the hash
-    const element = document.getElementById(id);
-    if (element) {
-      let parentDetails = element.parentNode;
-      while (parentDetails && parentDetails.tagName !== "DETAILS") {
-        parentDetails = parentDetails.parentNode;
-      }
-      if (parentDetails) {
-        parentDetails.setAttribute("open", "");
-        element.scrollIntoView(); // This is necessary to scroll to IDs inside collapsed `details` elements over a slow connection
-      }
-    }
-  }
-}
+import { openDetails } from "../../components/openDetails";
 
 function CssProperties() {
   // This prevents *other* sections closing when the URL hash changes on hash-link clicks and the page re-renders:
